Rename module from pool to sql
Renamed the database access module from 'pool' to 'sql' in multiple files throughout the application to enhance clarity. The name 'sql' better reflects the functionality of the module, which is about handling SQL database transactions, and makes the codebase more readable, which will help in future development and maintenance processes.
