+++ src/cmd/general_module/function/mod.rs
@@ -2,6 +2,6 @@ pub mod differed_response;
 pub mod get_guild_langage;
 pub mod html_parser;
 pub mod in_progress;
-pub mod pool;
 pub mod request;
+pub mod sql;
 pub mod trim;
